api-service/pyproject.toml
[project]
name = "api-service"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Leon" }]
dependencies = [
    "fastapi[standard]>=0.119.0",
    "sqlmodel>=0.0.27",
    "pydantic-settings>=2.11.0",
    "loguru>=0.7.3",
    "pathvalidate>=3.3.1",
    "aiobotocore>=2.25.0",
    "python-ulid>=3.1.0",
    "alembic>=1.17.0",
    "psycopg[binary]>=3.2",
    "pwdlib[argon2]>=0.2.1",
    "pyjwt[crypto]>=2.10.1",
    "rq>=2.6.0",
]
requires-python = ">=3.12"
readme = "README.md"


[project.optional-dependencies]
dev = [
    "ipykernel>=7.0",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "mimesis>=18.0.0",
    "types-aiobotocore[s3]>=2.25.0",
]


[tool.pdm]
distribution = false


[tool.pytest.ini_options]
log_level = "INFO"
asyncio_default_fixture_loop_scope = "function"
addopts = """
--exitfirst
--pdb
--pdbcls=IPython.terminal.debugger:TerminalPdb
-v
-r a
--cov
--cov-report=term
--cov-report=html
--cov-report=xml
--no-cov-on-fail
"""
